Convert from sqlite to postgres
Rewrite db wrapper

Update existing queries

+++ jonxhikari/data/static/build.sql
@@ -1,21 +1,23 @@
 CREATE TABLE IF NOT EXISTS guilds (
-    GuildID integer NOT NULL PRIMARY KEY,
-    Prefix text DEFAULT "$",
-    StarChannel integer DEFAULT 0
+    GuildID bigint NOT NULL PRIMARY KEY,
+    Prefix text DEFAULT '$',
+    StarChannel bigint DEFAULT 0
 );
 
 CREATE TABLE IF NOT EXISTS tags (
-    GuildID integer, TagOwner integer,
-    TagName text, TagContent text,
-    Uses integer DEFAULT 0,
+    GuildID bigint,
+    TagOwner bigint,
+    TagName text,
+    TagContent text,
+    Uses bigint DEFAULT 0,
     PRIMARY KEY (GuildID, TagName)
 );
 
 CREATE TABLE IF NOT EXISTS starboard (
-    BaseMessageID integer,
-    GuildID integer,
-    StarMessageID integer,
-    Stars integer DEFAULT 1,
+    BaseMessageID bigint,
+    GuildID bigint,
+    StarMessageID bigint,
+    Stars bigint DEFAULT 1,
 	FOREIGN KEY (GuildID) REFERENCES guilds(GuildID),
     PRIMARY KEY (BaseMessageID, GuildID)
 );
